apps/pyproject.toml
[tool.poetry]
name = "ninechronicles-seasonpass-worker"
version = "1.3.0"
description = "Season pass service of Nine Chronicles"
authors = ["hyeon <hyeon@planetariumhq.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
gql = { extras = ["all"], version = "^3.4.1" }
requests = "^2.31.0"
sqlalchemy = "^2.0.22"
psycopg2-binary = "^2.9.9"
pydantic = "^2.10"
pydantic-settings = "^2.8.1"
eth-utils = "^5.1.0"
eth-account = "^0.13.4"
pyasn1 = "^0.6.1"
pycryptodome = "^3.19.0"
hexbytes = "^1.2.1"
bencodex = "^1.0.1"
pyjwt = "^2.8.0"
pre-commit = "^4.1.0"
pika = "^1.3.2"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.28.65"
pytest = "^7.4.2"
ipython = "^8.16.1"
alembic = "^1.12.0"
aws-cdk-lib = "2.166.0"
constructs = "^10.3.0"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.106.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
